Au3AudioMeter adapter class
