+++ src/auaudio/CMakeLists.txt
@@ -8,10 +8,15 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/auaudiomodule.cpp
     ${CMAKE_CURRENT_LIST_DIR}/auaudiomodule.h
     ${CMAKE_CURRENT_LIST_DIR}/auaudiotypes.h
+    ${CMAKE_CURRENT_LIST_DIR}/iaudiometer.h
 
+    ${CMAKE_CURRENT_LIST_DIR}/internal/audiometer.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/audiometer.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/auqttimer.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/auqttimer.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/itimer.h
 )
+
+set(MODULE_LINK au3wrap) # For LockFreeQueue
 
 setup_module()
