+++ src/auaudio/iaudiometer.h
@@ -1,14 +1,24 @@
 /*
-* Audacity: A Digital Audio Editor
-*/
+ * Audacity: A Digital Audio Editor
+ */
 #pragma once
 
+#include "auaudio/auaudiotypes.h"
+
+#include "framework/global/async/channel.h"
+
+#include <chrono>
+#include <memory>
+#include <optional>
+
+constexpr int64_t MASTER_TRACK_ID = -2;
 using TimePoint = std::chrono::steady_clock::time_point;
 
-class AUDIO_DEVICES_API IMeterSender
+namespace au::auaudio {
+class IAudioMeter
 {
 public:
-    static constexpr int64_t MASTER_TRACK_ID = -2;
+    static std::unique_ptr<IAudioMeter> create();
 
     struct InterleavedSampleData
     {
@@ -38,10 +48,14 @@ public:
         bool operator<(const TrackId& other) const { return value < other.value; }
     };
 
-    virtual ~IMeterSender() = default;
+    using OptionalTimePoint = std::optional<TimePoint>;
+
+    virtual ~IAudioMeter() = default;
     virtual void push(uint8_t channel, const InterleavedSampleData& sampleData, TrackId = TrackId { MASTER_TRACK_ID }) = 0;
     virtual void start(double sampleRate) = 0;
     virtual void stop() = 0;
+    virtual muse::async::Channel<auaudio::audioch_t, auaudio::MeterSignal> dataChanged(TrackId key = TrackId { MASTER_TRACK_ID }) = 0;
 };
 
-using IMeterSenderPtr = std::shared_ptr<IMeterSender>;
+using IAudioMeterPtr = std::shared_ptr<IAudioMeter>;
+}
